cleanup: abstract group dialogs, change comments to messages, fix deprecated props
